bugfix: fix vfs dirent type
Issue: https://gitblueos.vivo.xyz/BlueOS/Kernel/BlueKernel/manifests/-/issues/89

+++ kernel/src/vfs/fatfs.rs
@@ -665,7 +665,12 @@ impl InodeOps for FatInode {
 
         // Handle special entries (., ..)
         if current_offset == 0 {
-            match reader.write_node(inner.attr.ino(), current_offset, inner.attr.type_(), ".") {
+            match reader.write_node(
+                inner.attr.ino(),
+                current_offset as i64,
+                inner.attr.type_(),
+                ".",
+            ) {
                 Ok(_) => {
                     count += 1;
                     current_offset += 1;
@@ -674,9 +679,12 @@ impl InodeOps for FatInode {
             }
         }
         if current_offset == 1 {
-            if let Err(e) =
-                reader.write_node(inner.attr.ino(), current_offset, inner.attr.type_(), "..")
-            {
+            if let Err(e) = reader.write_node(
+                inner.attr.ino(),
+                current_offset as i64,
+                inner.attr.type_(),
+                "..",
+            ) {
                 if count == 0 {
                     return Err(e);
                 }
@@ -689,7 +697,7 @@ impl InodeOps for FatInode {
         for (name, inode) in dir.children.iter().skip(start_idx) {
             match reader.write_node(
                 inode.inode_attr().ino(),
-                current_offset,
+                current_offset as i64,
                 inode.inode_attr().type_(),
                 name,
             ) {
