+++ kernel/src/vfs/tmpfs.rs
@@ -512,7 +512,12 @@ impl InodeOps for TmpInode {
 
         // Handle special entries (., ..)
         if current_offset == 0 {
-            match reader.write_node(inner.attr.ino(), current_offset, inner.attr.type_(), ".") {
+            match reader.write_node(
+                inner.attr.ino(),
+                current_offset as i64,
+                inner.attr.type_(),
+                ".",
+            ) {
                 Ok(_) => {
                     count += 1;
                     current_offset += 1;
@@ -522,9 +527,12 @@ impl InodeOps for TmpInode {
         }
 
         if current_offset == 1 {
-            if let Err(e) =
-                reader.write_node(inner.attr.ino(), current_offset, inner.attr.type_(), "..")
-            {
+            if let Err(e) = reader.write_node(
+                inner.attr.ino(),
+                current_offset as i64,
+                inner.attr.type_(),
+                "..",
+            ) {
                 if count == 0 {
                     return Err(e);
                 }
@@ -537,7 +545,7 @@ impl InodeOps for TmpInode {
         let start_idx = current_offset.saturating_sub(2);
         for (name, inode) in dir.children.iter().skip(start_idx) {
             let attr = inode.inode_attr();
-            match reader.write_node(attr.ino(), current_offset, attr.type_(), name) {
+            match reader.write_node(attr.ino(), current_offset as i64, attr.type_(), name) {
                 Ok(_) => {
                     count += 1;
                     current_offset += 1;
